+++ gpe/src/physics_debug_system.c
@@ -35,7 +35,7 @@ void reshape(int width, int height)
 {
 	glViewport(0, 0, width, height);
 	
-	double scale = cpfmin(width/640.0, height/480.0);
+	GLfloat scale = (GLfloat)cpfmin(width/640.0, height/480.0);
 	double hw = width*(0.5/scale);
 	double hh = height*(0.5/scale);
